user authentication and validation by admin implemented

+++ src/routes/entities.js
@@ -25,6 +25,7 @@ import {
 } from "../models/mysqlModels.js";
 import {FileContainer} from "../models/mongoDbModels.js";
 import {BadRequest, Unauthorized} from "../models/validation/errors.js";
+import {sendBasicEmail} from "../mailjet/sendEmail.js";
 import {validateIntegerId} from "../models/validation/utilityFunctions.js";
 
 const router = express.Router();
@@ -99,11 +100,33 @@ const model = (entity) => {
       };
   }
 };
+function userIsAdmin(req) {
+  if (!req.user || (req.user && req.user.role !== "admin"))
+    return [false, new Unauthorized("'admin' privileges required !")];
+  return [true, null];
+}
+function userIsOwnerOrAdmin(req, data) {
+  if (!req.user)
+    return [false, new Unauthorized("User must be authenticated !")];
+  if (req.user.role === "admin") return [true, null];
+  if (data.email !== req.user.email)
+    return [
+      false,
+      new Unauthorized(
+        "User account can only be accessed, updated or deleted by the account owner !"
+      ),
+    ];
+  return [true, null];
+}
 router.get(
   "/:model",
   routeHandler(async (req, res) => {
     const mdl = model(req.params.model);
-    const data = await mdl.model.findAll();
+    if (mdl.model === User) {
+      const cond = userIsAdmin(req);
+      if (!cond[0]) return res.send(cond[1]);
+    }
+    const data = await mdl.model.findAll({attributes: {exclude: ["pwd"]}});
     res.send({
       statusCode: "200",
       data,
@@ -120,6 +143,11 @@ router.get(
     const data = await mdl.model.findByPk(id);
     if (!data)
       return res.send(new BadRequest(`Record with id:${id} not found.`));
+    if (mdl.model === User) {
+      const cond = userIsOwnerOrAdmin(req, data);
+      if (!cond[0]) return res.send(cond[1]);
+      data.pwd = undefined;
+    }
     res.send({
       statusCode: "200",
       data,
@@ -131,6 +159,10 @@ router.post(
   authHandler,
   routeHandler(async (req, res) => {
     const mdl = model(req.params.model);
+    if (mdl.model === User) {
+      const cond = userIsAdmin(req);
+      if (!cond[0]) return res.send(cond[1]);
+    } //user is normally created through the register route
     const {error} = mdl.validate(req.body, "post");
     if (error) return res.send(new BadRequest(error.details[0].message));
     let data = null;
@@ -146,6 +178,7 @@ router.post(
         );
     }
     data = await mdl.model.create(req.body);
+    if (mdl.model === User) data.pwd = undefined;
     res.send({
       status: "OK",
       message: `Record successfully created with id:${data.id}.`,
@@ -166,7 +199,24 @@ router.patch(
       return res.send(new BadRequest(`Record with id:${id} not found.`));
     error = mdl.validate(req.body, "patch").error;
     if (error) return res.send(new BadRequest(error.details[0].message));
+    let userValidation = null;
+    if (mdl.model === User) {
+      const cond = userIsOwnerOrAdmin(req, data);
+      if (!cond[0]) return res.send(cond[1]);
+      userValidation = Object.keys(req.body).indexOf("validated") !== -1;
+      if (userValidation && req.user.role !== "admin")
+        return res.send(
+          new Unauthorized(`'admin' privileges required for user validation !`)
+        );
+    }
     await data.update(req.body);
+    if (userValidation && req.body.validated)
+      sendBasicEmail(
+        data.email,
+        "nationsounds_api: validation de votre compte",
+        `<b>${data.email}</b> (id:${data.id}) avec le rôle '${data.role}' a été validé.`
+      );
+    if (mdl.model === User) data.pwd = undefined;
     res.send({
       status: "OK",
       message: `Record successfully updated.`,
@@ -197,6 +247,10 @@ router.delete(
           )
         );
     }
+    if (mdl.model === User) {
+      const cond = userIsOwnerOrAdmin(req, data);
+      if (!cond[0]) return res.send(cond[1]);
+    }
     if (data.files_id) {
       const container = await FileContainer.findByIdAndDelete({
         _id: data.files_id,
@@ -207,6 +261,7 @@ router.delete(
         );
     }
     await data.destroy();
+    if (mdl.model === User) data.pwd = undefined;
     res.send({
       status: "OK",
       message: `Record '${id}' and associated files (if any) successfully deleted.`,
